Extract dxgi conversion functions into auxil
