+++ wgpu-hal/Cargo.toml
@@ -16,7 +16,7 @@ default = []
 metal = ["naga/msl-out", "block", "foreign-types"]
 vulkan = ["naga/spv-out", "ash", "gpu-alloc", "gpu-descriptor", "libloading", "inplace_it"]
 gles = ["naga/glsl-out", "glow", "egl", "libloading"]
-dx11 = ["naga/hlsl-out", "winapi/d3d11", "winapi/d3d11_1", "winapi/d3d11sdklayers", "winapi/dxgi1_6"]
+dx11 = ["naga/hlsl-out", "native", "winapi/d3d11", "winapi/d3d11_1", "winapi/d3d11sdklayers", "winapi/dxgi1_6"]
 dx12 = ["naga/hlsl-out", "native", "bit-set", "range-alloc", "winapi/d3d12", "winapi/d3d12shader", "winapi/d3d12sdklayers", "winapi/dxgi1_6"]
 renderdoc = ["libloading", "renderdoc-sys"]
 emscripten = ["gles"]
